commit
Starts implementation of get_char procedure. which utilizes the after_char procedure which will be implemented later. it places the current char being analyzed into register for the argument then stores 59 into t1 which corresponds to the decimal value for a semicolon. then will take the current char and see if it is equal to a semicolon and if it is will call the again procedure not yet implemented to clear out the stack and increment the counter