Commit
Starts implementation of the checker procedure which checks the char to see if its an uppercase letter a lower case letter or a number is going to need 3 other procedures for adjust those values corresponding. s7is initialized as 65 to check to see if its an uppercase letter. Similar to project 2 and 1