Commit
Finished implementation for finished procedure. t8 is initialized as the count and the return value is loaded in v0. less than instruction is used for a3 and then it goes to the after procedure which will be implemented later. Also need a procedure to multiply by the base to get the base conversion that was initialized at the beginning of the program